delete trash/ move some components
