+++ components/common/EducationCard/index.ts
