+++ components/common/EducationCard/types.ts
