--- components/common/TestComponent/TestComponent.module.css
@@ -1,6 +1,0 @@
-.test-background {
-  background-color: #f4f4f4;
-  padding: 20px;
-  border-radius: 5px;
-  margin-bottom: 20px;
-}
