--- components/common/TestComponent/index.ts
@@ -1,1 +1,0 @@
-export * from './TestComponent';
