--- components/common/TestComponent/types.ts
@@ -1,1 +1,0 @@
-// типи, які відносяться до даного компонента
